+++ SleepModel/Sense-Bridging-Header.h
@@ -33,3 +33,4 @@
 
 #import "HEMAppUsage.h"
 #import "HEMStyle.h"
+#import "HEMPresenter.h"